+++ JuliaBUGS/src/compiler_pass.jl
@@ -711,7 +711,7 @@ function build_node_functions(
                 statement.args[2], statement.args[3]
             end
             args, node_func_expr = make_function_expr(lhs, rhs, eval_env)
-            node_func = Core.eval(eval_module, node_func_expr)
+            node_func = _make_misty_closure(node_func_expr, eval_module)
             f_dict[statement] = (args, node_func_expr, node_func)
         elseif Meta.isexpr(statement, :for)
             loop_var, _, _, body = decompose_for_expr(statement)